Rename aca_require_temp to aca_lts_eval.py
